CodeMirror 6 preview (first draft) (#490)
