+++ source/WebApp/WebApp.csproj
@@ -34,17 +34,6 @@
   <ItemGroup Condition="'@(_StaticOutputs)' == ''">
     <_StaticOutputs Include="$(wwwroot)\.any" />
   </ItemGroup>
-  <Target Name="NpmInstall" BeforeTargets="NpmBuild" Inputs="package.json" Outputs="!package.json.installed">
-    <Exec Command="npm install" EnvironmentVariables="NODE_ENV=" />
-    <Copy SourceFiles="package.json" DestinationFiles="!package.json.installed" />
-  </Target>
-  <Target Name="NpmBuild" AfterTargets="Build" Inputs="@(_StaticInputs)" Outputs="@(_StaticOutputs)">
-    <Delete Files="@(_StaticOutputs)" />
-    <Exec Command="npm run build" />
-    <ItemGroup>
-      <Content Include="wwwroot\**\*.*" />
-    </ItemGroup>
-  </Target>
 
   <Target Name="CopyFrameworkXmlDocumentation" AfterTargets="ResolveReferences">
     <ItemGroup>
