+++ source/WebApp/components/app-cm6-preview-manager.html
@@ -1,0 +1,3 @@
+<div>
+  Editor: <button v-on:click="toggle">{{editorName}}</button>
+</div>
