+++ source/WebApp/components/app-code-edit.html
@@ -1,0 +1,1 @@
+<component v-bind:is="editor" v-bind="$props" v-on="$listeners"></component>
