+++ source/WebApp/index.html
@@ -164,6 +164,7 @@
       <portal-target name="modals" multiple slim></portal-target>
     </main>
     <footer>
+      <app-cm6-preview-manager></app-cm6-preview-manager>
       <app-theme-manager></app-theme-manager>
       <span>
         Built by Andrey Shchekin (<a href="http://twitter.com/ashmind">@ashmind</a>) — see <a href="http://github.com/ashmind/SharpLab">SharpLab on GitHub</a>.
