+++ source/WebApp/package.json
@@ -73,7 +73,7 @@
     "postcss": "7.0.2",
     "postcss-csso": "3.0.0",
     "puppeteer": "2.1.1",
-    "rollup": "2.3.2",
+    "rollup": "2.9.1",
     "rollup-plugin-babel": "4.4.0",
     "rollup-plugin-commonjs": "10.1.0",
     "rollup-plugin-node-resolve": "3.0.0",
@@ -94,6 +94,7 @@
     "lz-string": "1.4.4",
     "marked": "0.7.0",
     "mirrorsharp": "2.0.0-preview-04",
+    "mirrorsharp-codemirror-6-preview": "0.0.1",
     "normalize.css": "8.0.1",
     "portal-vue": "2.1.7",
     "resize-observer-polyfill": "1.5.1",
